# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Restaurant/Bar API
  version: 1.0.0
  description: >
    Backend para gestión de restaurante/bar con multisucursal.
    Incluye auth con JWT + Refresh, roles (MANAGER, ADMIN, WAITER, CASHIER),
    Branches, Tables, Products, Orders (con items), Stock y reportes de stock.
servers:
  - url: http://localhost:4000
    description: Local
...
paths:
  /api/auth/register:
    post:
      tags: [Auth]
      summary: Register user
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RegisterRequest' }
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserPublic' }
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Login (returns access + refresh)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        "200":
          description: Login OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LoginResponse' }
  /api/auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh session (rotate refresh, new access token)
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken:
                  type: string
                  description: Optional if refreshToken cookie is present
      responses:
        "200":
          description: New tokens
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RefreshResponse' }
...
  /api/auth/profile:
    get:
      tags: [Auth]
      summary: Get authenticated profile
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserPublic' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    RegisterRequest:
      type: object
      required: [name, email, password, role]
      properties:
        name: { type: string }
        email: { type: string, format: email }
        password: { type: string, format: password }
        role: { type: string, enum: [MANAGER, ADMIN, WAITER, CASHIER] }
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, format: password }
    LoginResponse:
      type: object
      properties:
        message: { type: string }
        accessToken: { type: string }
        refreshToken: { type: string }
        user: { $ref: '#/components/schemas/UserPublic' }
    RefreshResponse:
      type: object
      properties:
        message: { type: string }
        accessToken: { type: string }
        refreshToken: { type: string }
    UserPublic:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        email: { type: string }
        role: { type: string }
        branchId: { type: integer, nullable: true }
...
security:
  - bearerAuth: []